requirements.txt
amqp==2.2.2
anyjson==0.3.3
appdirs==1.4.3
attrs==18.1.0
billiard==3.5.0.3
cached-property==1.4.2
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
circus==0.15.0
click==6.7
discogs-client==2.2.1
dj-database-url==0.5.0
Django==1.11.11
django-celery==3.2.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-el-pagination==3.2.4
django-filter==1.1.0
django-widget-tweaks==1.4.2
djangorestframework==3.7.7
docopt==0.6.2
gunicorn==19.9.0
idna==2.6
jsonschema==2.6.0
kombu==4.1.0
oauthlib==2.0.7
Pillow==5.1.0
psutil==5.4.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-mpd2==1.0.0
pytz==2018.3
PyYAML==3.12
pyzmq==16.0.4
redis==2.10.6
requests==2.18.4
six==1.11.0
texttable==0.9.1
tornado==4.5.3
urllib3==1.22
vine==1.1.4
websocket-client==0.47.0
